requirements.txt
fastapi
uvicorn[standard]
anthropic
gitpython
pylint
bandit
radon
pytest
python-dotenv
pydantic
aiofiles
